002: changes to cron script

+++ scripts/create-crons.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
 
 #temp cron
-crontab -l > temp_cron
-echo "*/1 * * * * sh $(pwd)/scripts/cron/tempcheck.sh" >> temp_cron
-#install new cron file
-crontab temp_cron
-rm temp_cron
+if crontab -l > temp_cron; then
+    #install new cron file
+    echo "*/1 * * * * sh $(pwd)/scripts/cron/tempcheck.sh" >> temp_cron
+    crontab temp_cron
+    rm temp_cron
+    echo -e "\033[0;36m Cron jobs added"
+else 
+    echo -e "\033[0;31m Error: Could not add cron jobs"
+fi
+
